[None] refactor: AuthService, MemberService Interface 삭제
